renamed contrib projects for akka cluster
