+++ src/Akka.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+# Visual Studio 14
+VisualStudioVersion = 14.0.22823.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{69279534-1DBA-4115-BF8B-03F77FC8125E}"
 EndProject
@@ -194,13 +194,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Persistence.SqlServer"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Persistence.SqlServer.Tests", "contrib\persistence\Akka.Persistence.SqlServer.Tests\Akka.Persistence.SqlServer.Tests.csproj", "{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ClusterSharding", "ClusterSharding", "{D8F27F25-4EC3-4746-B3A8-BC94413A6830}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Cluster", "Cluster", "{52999BCC-7C31-4D38-92E9-B564077294A9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.ClusterSharding", "contrib\clustersharding\Akka.ClusterSharding\Akka.ClusterSharding.csproj", "{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Cluster.Sharding", "contrib\cluster\Akka.Cluster.Sharding\Akka.Cluster.Sharding.csproj", "{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ClusterSingleton", "ClusterSingleton", "{701D29CE-E37B-4C23-944B-55E873093B49}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.ClusterSingleton", "contrib\clustersingleton\Akka.ClusterSingleton\Akka.ClusterSingleton.csproj", "{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Cluster.Tools", "contrib\cluster\Akka.Cluster.Tools\Akka.Cluster.Tools.csproj", "{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -711,22 +709,22 @@ Global
 		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Release Mono|Any CPU.Build.0 = Release|Any CPU
 		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug Mono|Any CPU.ActiveCfg = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug Mono|Any CPU.Build.0 = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release Mono|Any CPU.ActiveCfg = Release|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release Mono|Any CPU.Build.0 = Release|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E}.Debug Mono|Any CPU.ActiveCfg = Debug|Any CPU
-		{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E}.Debug Mono|Any CPU.Build.0 = Debug|Any CPU
-		{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E}.Release Mono|Any CPU.ActiveCfg = Release|Any CPU
-		{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E}.Release Mono|Any CPU.Build.0 = Release|Any CPU
-		{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421}.Debug Mono|Any CPU.ActiveCfg = Debug|Any CPU
+		{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421}.Debug Mono|Any CPU.Build.0 = Debug|Any CPU
+		{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421}.Release Mono|Any CPU.ActiveCfg = Release|Any CPU
+		{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421}.Release Mono|Any CPU.Build.0 = Release|Any CPU
+		{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E}.Debug Mono|Any CPU.ActiveCfg = Debug|Any CPU
+		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E}.Debug Mono|Any CPU.Build.0 = Debug|Any CPU
+		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E}.Release Mono|Any CPU.ActiveCfg = Release|Any CPU
+		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E}.Release Mono|Any CPU.Build.0 = Release|Any CPU
+		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -813,9 +811,8 @@ Global
 		{264C22A4-CAFC-41F6-B82C-4DDC5C196767} = {588C1513-FAB6-42C3-B6FC-3485F13620CF}
 		{BAC85686-AFC4-413E-98DC-5ED8F468BC63} = {264C22A4-CAFC-41F6-B82C-4DDC5C196767}
 		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE} = {264C22A4-CAFC-41F6-B82C-4DDC5C196767}
-		{D8F27F25-4EC3-4746-B3A8-BC94413A6830} = {588C1513-FAB6-42C3-B6FC-3485F13620CF}
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4} = {D8F27F25-4EC3-4746-B3A8-BC94413A6830}
-		{701D29CE-E37B-4C23-944B-55E873093B49} = {588C1513-FAB6-42C3-B6FC-3485F13620CF}
-		{3644B55E-7A7B-4402-89FB-1EF1F5EE6A8E} = {701D29CE-E37B-4C23-944B-55E873093B49}
+		{52999BCC-7C31-4D38-92E9-B564077294A9} = {588C1513-FAB6-42C3-B6FC-3485F13620CF}
+		{2CB7BC70-DEEB-4DA2-9F61-3F9ED6373421} = {52999BCC-7C31-4D38-92E9-B564077294A9}
+		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E} = {52999BCC-7C31-4D38-92E9-B564077294A9}
 	EndGlobalSection
 EndGlobal
