+++ src/contrib/cluster/Akka.Cluster.Sharding/ClusterShardingExtension.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Akka.Actor;
+﻿using Akka.Actor;
 
-namespace Akka.ClusterSharding
+namespace Akka.Cluster.Sharding
 {
     public class ClusterShardingExtension : ExtensionIdProvider<ClusterSharding>
     {
