+++ src/contrib/cluster/Akka.Cluster.Sharding/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Bcl.Immutable" version="1.0.34" targetFramework="net45" userInstalled="true" />
+</packages>
