--- src/contrib/clustersharding/Akka.ClusterSharding/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Bcl.Immutable" version="1.0.34" targetFramework="net45" />
-</packages>
